Fix JRE 6 compatibility, mark README with warning

+++ src/com/rusticisoftware/hostedengine/client/ServiceRequest.java
@@ -387,7 +387,7 @@ public class ServiceRequest {
 
         final String introStr = intro.toString();
         final String outroStr = outro.toString();
-        long requestLength = introStr.getBytes("utf8").length + file.length() + outroStr.getBytes("utf8").length;
+        int requestLength = (int)(introStr.getBytes("utf8").length + file.length() + outroStr.getBytes("utf8").length);
 
         connection.setDoOutput(true);
         connection.setRequestProperty("Content-Type", "multipart/form-data; boundary=" + boundary);
